Production-ready improvements and warehouse UX enhancements
MAJOR IMPROVEMENTS:
==================

Bug Fixes:
- Fixed receipt deletion with proper transaction reversal and batch cleanup
- Fixed float precision issues for monetary values (Numeric(10,4) instead of Float)
- Added comprehensive error handling and logging

Security Enhancements:
- Added CSRF protection to all forms with Flask-WTF
- Implemented environment-based configuration with python-dotenv
- Added .gitignore to protect sensitive data (.env, database, uploads)
- Moved SECRET_KEY to environment variables
- Added rotating file logging for audit trail
- Non-root Docker container for security

Warehouse Worker UX Optimizations:
- Added large, touch-friendly buttons (btn-lg, btn-xl classes)
- Implemented barcode scanner auto-detection and processing
- Added keyboard shortcuts (N=New, R=Receipt, T=Transfer, A=Adjustment, S=Scrap, /=Search, ?=Help)
- Implemented confirmation dialogs for destructive actions (delete operations)
- Added visual feedback animations (success-state, error-state, shake, pulse)
- Auto-focus first input field on page load
- Enter key navigation between form fields
- Enhanced CSS for warehouse mode with high-contrast options

Production Deployment:
- Created Dockerfile with multi-stage build
- Added docker-compose.yml for easy deployment
- Created deploy.sh script for automated deployment (Docker or systemd)
- Added .dockerignore for optimized builds
- Implemented /health endpoint for monitoring and load balancers
- Added comprehensive DEPLOYMENT.md guide
- Created README.md with full documentation

Technical Improvements:
- Added python-dotenv for environment variable management
- Added Flask-WTF for CSRF and form validation
- Added requests library for health checks
- Configured gunicorn for production server
- Added rotating log handler (10MB max, 10 backups)
- Database models now use Numeric for all cost_per_unit fields
- Improved error messages and user feedback

Files Changed:
==============
NEW:
- .gitignore
- .env.example
- .dockerignore
- Dockerfile
- docker-compose.yml
- deploy.sh
- README.md
- DEPLOYMENT.md
- inventory-management/static/js/warehouse.js

MODIFIED:
- inventory-management/app.py (CSRF, logging, health check)
- inventory-management/config.py (environment variables, dotenv)
- inventory-management/models.py (Numeric for monetary values)
- inventory-management/requirements.txt (Flask-WTF, python-dotenv, requests)
- inventory-management/routes/receipts.py (fixed deletion)
- inventory-management/static/css/style.css (warehouse UX enhancements)
- inventory-management/templates/base.html (warehouse.js script)

DEPLOYMENT READY:
- Docker: ./deploy.sh -> Option 1
- Systemd: ./deploy.sh -> Option 2
- Default credentials: admin/admin123 (CHANGE IMMEDIATELY)

+++ inventory-management/models.py
@@ -193,7 +193,7 @@ class Batch(db.Model):
     quantity_available = db.Column(db.Float, nullable=False)
 
     received_date = db.Column(db.DateTime, nullable=False, default=datetime.utcnow, index=True)
-    cost_per_unit = db.Column(db.Float, nullable=False)
+    cost_per_unit = db.Column(db.Numeric(10, 4), nullable=False)  # Precision for currency
     supplier_batch_number = db.Column(db.String(100))
     po_number = db.Column(db.String(100))  # External ERP reference
 
@@ -276,7 +276,7 @@ class ReceiptItem(db.Model):
     bin_id = db.Column(db.Integer, db.ForeignKey('bins.id'), nullable=True)
 
     quantity = db.Column(db.Float, nullable=False)
-    cost_per_unit = db.Column(db.Float, nullable=False)
+    cost_per_unit = db.Column(db.Numeric(10, 4), nullable=False)  # Precision for currency
     supplier_batch_number = db.Column(db.String(100))
     batch_id = db.Column(db.Integer, db.ForeignKey('batches.id'), nullable=True)  # Created batch
 
@@ -367,7 +367,7 @@ class TransferBatch(db.Model):
     transfer_id = db.Column(db.Integer, db.ForeignKey('transfers.id'), nullable=False)
     batch_id = db.Column(db.Integer, db.ForeignKey('batches.id'), nullable=False)
     quantity_transferred = db.Column(db.Float, nullable=False)
-    cost_per_unit = db.Column(db.Float, nullable=False)  # From batch at time of transfer
+    cost_per_unit = db.Column(db.Numeric(10, 4), nullable=False)  # From batch at time of transfer
 
     def __repr__(self):
         return f'<TransferBatch {self.transfer_id}: Batch {self.batch_id}, Qty {self.quantity_transferred}>'
@@ -455,7 +455,7 @@ class ScrapBatch(db.Model):
     scrap_id = db.Column(db.Integer, db.ForeignKey('scraps.id'), nullable=False)
     batch_id = db.Column(db.Integer, db.ForeignKey('batches.id'), nullable=False)
     quantity_scrapped = db.Column(db.Float, nullable=False)
-    cost_per_unit = db.Column(db.Float, nullable=False)  # From batch at time of scrap
+    cost_per_unit = db.Column(db.Numeric(10, 4), nullable=False)  # From batch at time of scrap
 
     def __repr__(self):
         return f'<ScrapBatch {self.scrap_id}: Batch {self.batch_id}, Qty {self.quantity_scrapped}>'
